+++ let-there-be-light-c/floodfill.h
@@ -2,6 +2,7 @@
 #define FLOODFILL_H
 
 #include <stdbool.h>
+#include <stddef.h>
 
 #include "config.h"
 #include "list.h"
@@ -19,12 +20,16 @@ typedef struct FloodState {
   FrontierList* frontiers;
   const Tile (*tiles)[MAP_SIZE];
   bool visited[MAP_SIZE][MAP_SIZE];
-  unsigned pathLength;
+  size_t pathLength;
 } FloodState;
 
-FloodState* floodGenerate(const Tile tiles[MAP_SIZE][MAP_SIZE], size_t x, size_t y);
+FloodState* floodGenerate(const Tile tiles[MAP_SIZE][MAP_SIZE],
+                          size_t x,
+                          size_t y);
 
-FloodState* floodFill(Tile tiles[MAP_SIZE][MAP_SIZE], size_t x, size_t y);
+FloodState* floodFill(Tile tiles[MAP_SIZE][MAP_SIZE],
+                      size_t x,
+                      size_t y);
 
 void floodDestory(FloodState* state);
