+++ let-there-be-light-c/maze.h
@@ -6,6 +6,9 @@
 #include "config.h"
 #include "tile.h"
 
-unsigned initMaze(size_t spawnerCount, size_t minDistance, size_t maxDistance, Tile tiles[MAP_SIZE][MAP_SIZE]);
+size_t initMaze(size_t spawnerCount,
+                size_t minDistance,
+                size_t maxDistance,
+                Tile tiles[MAP_SIZE][MAP_SIZE]);
 
 #endif
